- updated the migration files - added repositories - added forms - add song is now working - fixed issue with renderError

+++ domain/user/user.go
@@ -15,7 +15,7 @@ import (
 const ProviderDiscord = "discord"
 
 type User struct {
-	StorageID         int
+	StorageID         int64
 	Username          string
 	Email             string
 	Name              string
